requirements.txt
djangorestframework
psycopg2-binary
psycopg2
decouple
Pillow
social-auth-app-django
